refactor: update navigation button labels and enhance stepper layout
- Changed the previous button label from "Précédent" to "Retour" for better clarity in navigation.
- Adjusted the layout spacing in StepperWrapper for improved visual consistency.
- Removed the deprecated StepFourForm, streamlining the project creation process.
- Introduced StepThreeForm to replace the removed form, integrating role management and tech stack selection.
