+++ apps/web/src/features/projects/forms/scratch/step-two-form.tsx
@@ -128,6 +128,7 @@ export function StepTwoForm() {
 
       <FormNavigationButtons
         onPrevious={handlePrevious}
+        previousLabel="Retour"
         nextLabel="Suivant"
         isLoading={isSubmitting}
         nextType="submit"
