+++ apps/web/src/features/projects/views/stepper-views/github/step-three.view.tsx
@@ -6,7 +6,7 @@ import { useState } from "react";
 import { Button } from "@/shared/components/ui/button";
 
 import { StepperWrapper } from "../../../components/stepper/stepper-wrapper.component";
-import { StepThreeForm } from "../../../forms/step-three.form";
+import { StepThreeForm } from "../../../forms/scratch/step-three.form";
 import { useProjectCreateStore } from "../../../stores/project-create.store";
 
 export default function StepThreeView() {
